fraud detection/backend/requirements.txt
neo4j>=5.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
xgboost>=2.0.0
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
joblib>=1.3.0
python-dotenv>=1.0.0
